+++ billy/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.9.4'
+__version__ = '0.9.5'
 
 from billy.conf import settings